Use cors - closes #38 - remove set headers and use calls - rename controller files
