+++ src/routes/owners.js
@@ -1,5 +1,5 @@
 import express from "express";
-import owners from "../controllers/OwnerController";
+import owners from "../controllers/ownerController";
 const route = express.Router();
 
 route.post("/owners", owners.createOwner);
